+++ trunk/xymond/xymond_history.c
@@ -156,7 +156,7 @@ int main(int argc, char *argv[])
 	{
 		FILE *pidfd = fopen(pidfn, "w");
 		if (pidfd) {
-			fprintf(pidfd, "%d\n", getpid());
+			fprintf(pidfd, "%lu\n", (unsigned long)getpid());
 			fclose(pidfd);
 		}
 	}
